// src/types/products.ts
export  interface product {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: string;
    availabilityStatus: string;
    thumbnail: string;
    quantity : number;
    
    
}